fix order in shard level
