+++ src/core/storage/ordered_storage.rs
@@ -106,7 +106,7 @@ impl OrderedStorage {
                         &mut tables,
                     );
 
-                    Self::merge_disk_tables(shards.clone(), metadata.clone(), storage_path.clone());
+                    // Self::merge_disk_tables(shards.clone(), metadata.clone(), storage_path.clone());
                     return;
                 }
