+++ tests/sync_table_test.rs
@@ -76,34 +76,3 @@ fn test_mpsc_table() -> io::Result<()> {
 
     Ok(())
 }
-
-#[test]
-fn test_override_entry() -> io::Result<()> {
-    let table_name = "/tmp/kvs/test/test_override_entry";
-    let config = StorageConfig::default_config();
-
-    let key = FlexibleField::new([1, 1, 1, 1]);
-
-    let value = FlexibleField::new([2, 2, 2, 2]);
-    let entry1 = FlexibleUserEntry::new(key.clone(), value);
-
-    let value = FlexibleField::new([4, 4, 4, 4]);
-    let entry2 = FlexibleUserEntry::new(key.clone(), value);
-
-    {
-        let table = OrderedStorage::new(table_name, config.clone());
-        table.put(&entry1).unwrap();
-    }
-
-    {
-        let table = OrderedStorage::new(table_name, config.clone());
-        table.put(&entry2).unwrap();
-    }
-
-    let table = OrderedStorage::new(table_name, config.clone());
-    let result = table.get(&key).unwrap().expect("value was inserted");
-
-    assert_eq!(result, *entry2.get_value());
-
-    Ok(())
-}
